refactor!(test-utils): remvoe test-secert-manager macros
