+++ oprf-test-utils/src/secret_manager/test_secret_manager.rs
@@ -13,108 +13,6 @@ use rand::{CryptoRng, Rng};
 
 use crate::TEST_TIMEOUT;
 
-#[macro_export]
-macro_rules! key_gen_test_secret_manager {
-    ($trait: path, $name: ident, $types:path, $dlog: path) => {
-        mod impl_key_gen_secret_manager {
-            use eyre::Context;
-            use $dlog;
-            use $types::{OprfKeyId, ShareEpoch, async_trait::async_trait, crypto::OprfPublicKey};
-
-            // need a new type to implement the trait
-            pub struct $name(pub std::sync::Arc<$crate::test_secret_manager::TestSecretManager>);
-
-            impl $name {
-                pub fn wallet_private_key_hex_string(&self) -> String {
-                    self.0.wallet_private_key_hex_string()
-                }
-            }
-
-            #[async_trait]
-            impl $trait for $name {
-                async fn store_wallet_address(&self, address: String) -> eyre::Result<()> {
-                    self.0.store_wallet_address(address).await
-                }
-                async fn ping(&self) -> eyre::Result<()> {
-                    // noop
-                    Ok(())
-                }
-
-                async fn get_share_by_epoch(
-                    &self,
-                    oprf_key_id: OprfKeyId,
-                    generated_epoch: ShareEpoch,
-                ) -> eyre::Result<Option<DLogShareShamir>> {
-                    self.0
-                        .get_share_by_epoch(oprf_key_id, generated_epoch)
-                        .await
-                }
-
-                async fn remove_oprf_key_material(&self, rp_id: OprfKeyId) -> eyre::Result<()> {
-                    self.0
-                        .remove_oprf_key_material(rp_id)
-                        .await
-                        .context("while remove oprf key material")?;
-                    Ok(())
-                }
-
-                async fn store_dlog_share(
-                    &self,
-                    oprf_key_id: OprfKeyId,
-                    public_key: OprfPublicKey,
-                    epoch: ShareEpoch,
-                    share: DLogShareShamir,
-                ) -> eyre::Result<()> {
-                    self.0
-                        .store_dlog_share(oprf_key_id, public_key, epoch, share)
-                        .await
-                        .context("while store DlogShare")?;
-                    Ok(())
-                }
-            }
-        }
-        use impl_key_gen_secret_manager::$name;
-    };
-}
-
-#[macro_export]
-macro_rules! oprf_node_test_secret_manager {
-    ($secret_manager: path, $name: ident, $types:path, $dlog: path) => {
-        mod impl_node_secret_manager {
-            use alloy::primitives::Address;
-            use $dlog;
-            use $types::{
-                OprfKeyId, ShareEpoch, async_trait::async_trait, crypto::OprfKeyMaterial,
-            };
-
-            // need a new type to implement the trait
-            pub struct $name(pub std::sync::Arc<$crate::test_secret_manager::TestSecretManager>);
-
-            #[async_trait]
-            impl $secret_manager for $name {
-                async fn load_address(&self) -> eyre::Result<Address> {
-                    self.0.load_address().await
-                }
-
-                async fn load_secrets(
-                    &self,
-                ) -> eyre::Result<std::collections::HashMap<OprfKeyId, OprfKeyMaterial>> {
-                    Ok(self.0.store.lock().clone())
-                }
-
-                async fn get_oprf_key_material(
-                    &self,
-                    oprf_key_id: OprfKeyId,
-                    epoch: ShareEpoch,
-                ) -> eyre::Result<Option<OprfKeyMaterial>> {
-                    self.0.get_oprf_key_material(oprf_key_id, epoch).await
-                }
-            }
-        }
-        use impl_node_secret_manager::$name;
-    };
-}
-
 #[derive(Clone)]
 pub struct TestSecretManager {
     pub wallet_private_key: PrivateKeySigner,
